ch(code-refactor): code fix
- folder restructuring
- code refactor

[finishes #169633257]
