+++ .eslintrc.json
@@ -17,6 +17,6 @@
     },
     "rules": {
         "no-console": "off",
-        "linebreak-style": "off"
+        "linebreak-style": "on"
     }
 }
