+++ server/helpers/authToken.js
