+++ server/routes/entriesRoute.js
@@ -6,18 +6,10 @@ import verifyToken from '../middlewares/verifyToken';
 const router = express.Router();
 
 
-/* GET all user entries */
 router.get('/entries', verifyToken, EntryController.getAllEntries);
-
-/* GET a single entry */
 router.get('/entries/:id', verifyToken,validate.paramValidation, EntryController.getEntry);
-
-/* POST a new entry */
 router.post('/entries/', verifyToken, validate.entry, EntryController.addEntry);
-
-/* PUT new a data in existing entry */
 router.patch('/entries/:id', verifyToken, validate.paramValidation, validate.entry, EntryController.modifyEntry);
-
 router.delete('/entries/:id', verifyToken, validate.paramValidation, EntryController.deleteEntry)
 
 export default router;
