+++ server/routes/userRoute.js
@@ -5,10 +5,7 @@ import validate from '../middlewares/validator';
 const router = express.Router();
 
 
-/* Create a new user */
 router.post('/signup', validate.signUp, UserController.createUser);
-
-/* Login user */
 router.post('/signin', validate.signIn, UserController.loginUser);
 
 export default router;
